process nid ocr

+++ app/src/main/java/com/sm/spagent/model/Ocr.kt
@@ -2,6 +2,6 @@ package com.sm.spagent.model
 
 data class Ocr(
   val imageFile: String,
-  val nid: Int?,
+  val nid: Long?,
   val dob: String?
 )
